add verbose system to avoid console.log

+++ server/rest/components(C-M-R)/block/model.js
@@ -10,9 +10,10 @@ class Block {
 
   async create(blockingUserId, blockedUserId) {
     try {
-      console.log(
-        `INSERT INTO public."Block" (blockingUser, blockedUser) VALUES (${blockingUserId}, ${blockedUserId} RETURNING id)`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `INSERT INTO public."Block" (blockingUser, blockedUser) VALUES (${blockingUserId}, ${blockedUserId} RETURNING id)`,
+        );
       return await db
         .any(
           'INSERT INTO public."Block" ("blockingUser", "blockedUser") VALUES ($1, $2) RETURNING id',
@@ -26,39 +27,44 @@ class Block {
           };
         });
     } catch (err) {
-      console.log(err, 'in model Block.create()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Block.create()');
       return { created: false, error: err };
     }
   }
 
   async exists(blockingUser, blockedUser) {
     try {
-      console.log(
-        `SELECT exists(SELECT from public."Block" WHERE "blockingUser" = ${blockingUser} AND "blockedUser" = ${blockedUser})`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT exists(SELECT from public."Block" WHERE "blockingUser" = ${blockingUser} AND "blockedUser" = ${blockedUser})`,
+        );
       const result = await db.any(
         `SELECT exists(SELECT from public."Block" WHERE "blockingUser" = $1 AND "blockedUser" = $2);`,
         [blockingUser, blockedUser],
       );
       return result[0].exists;
     } catch (err) {
-      console.log(err, 'in model Block.exists()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Block.exists()');
       return null;
     }
   }
 
   async delete(blockingUser, blockedUser) {
     try {
-      console.log(
-        `DELETE FROM public."Block" WHERE "blockingUser" = ${blockingUser} AND "blockedUser" = ${blockedUser}`,
-      );
-      const result = await db.any(
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `DELETE FROM public."Block" WHERE "blockingUser" = ${blockingUser} AND "blockedUser" = ${blockedUser}`,
+        );
+      await db.any(
         'DELETE FROM public."Block" WHERE "blockingUser" = $1  AND "blockedUser" = $2 ',
         [blockingUser, blockedUser],
       );
       return { success: true, deleted: true };
     } catch (err) {
-      console.log(err, 'in model User.delete()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model User.delete()');
       return { deleted: false, error: err };
     }
   }
