+++ server/rest/components(C-M-R)/like/model.js
@@ -10,9 +10,10 @@ class Like {
 
   async create(likingUserId, likedUserId) {
     try {
-      console.log(
-        `INSERT INTO public."Like" (likingUser, likedUser, date) VALUES (${likingUserId}, ${likedUserId}, Now() RETURNING id)`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `INSERT INTO public."Like" (likingUser, likedUser, date) VALUES (${likingUserId}, ${likedUserId}, Now() RETURNING id)`,
+        );
       return await db
         .any(
           'INSERT INTO public."Like" ("likingUser", "likedUser", date) VALUES ($1, $2, NOW()) RETURNING id, EXISTS(SELECT * FROM public."Like" WHERE "likingUser" = $2 AND "likedUser" = $1) AS match',
@@ -27,7 +28,8 @@ class Like {
           };
         });
     } catch (err) {
-      console.log(err, 'in model Like.create()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.create()');
       return { created: false, error: err };
     }
   }
@@ -35,14 +37,16 @@ class Like {
   async getBy(type, value) {
     try {
       if (!this.isValidType(type)) {
-        console.log(`Like.getBy(): ${type} is not an authorized type`);
+        if (process.env.VERBOSE === 'true')
+          console.log(`Like.getBy(): ${type} is not an authorized type`);
         return null;
       }
-      console.log(
-        `SELECT firstname, username, birthDate, location, popularityRate, profilePicture, date
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT firstname, username, birthDate, location, popularityRate, profilePicture, date
         FROM public."Like"
         WHERE ${type} = ${value}`,
-      );
+        );
       const result = await db.any(
         `SELECT firstname, username, "birthDate", location, "popularityRate", "profilePicture", "likingUser" AS visitor, "likedUser" AS visited, date,
         EXISTS(SELECT * FROM public."Like" AS secondlike WHERE secondlike."likingUser" = $2 AND secondlike."likedUser" = "Like"."likingUser") AS liking
@@ -62,43 +66,49 @@ class Like {
       });
       return result;
     } catch (err) {
-      console.log(err, 'in model Like.getBy()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.getBy()');
       return null;
     }
   }
 
   async getAll() {
     try {
-      console.log('SELECT * FROM public."Like"');
+      if (process.env.VERBOSE === 'true')
+        console.log('SELECT * FROM public."Like"');
       const result = await db.any('SELECT * FROM public."Like"');
       return result;
     } catch (err) {
-      console.log(err, 'in model Like.getAll()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.getAll()');
       return null;
     }
   }
 
   async exists(likingUser, likedUser) {
     try {
-      console.log(
-        `SELECT exists(SELECT from public."Like" WHERE "likingUser" = ${likingUser} AND "likedUser" = ${likedUser})`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT exists(SELECT from public."Like" WHERE "likingUser" = ${likingUser} AND "likedUser" = ${likedUser})`,
+        );
       const result = await db.any(
         `SELECT exists(SELECT from public."Like" WHERE "likingUser" = $1 AND "likedUser" = $2);`,
         [likingUser, likedUser],
       );
       return result[0].exists;
     } catch (err) {
-      console.log(err, 'in model Like.exists()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.exists()');
       return null;
     }
   }
 
   async relationship(visitorUser, visitedUser) {
     try {
-      console.log(
-        `SELECT exists(SELECT from public."Like" WHERE "likingUser" = $1 AND "likedUser" = $2) AS visitorlikevisited, exists(SELECT from public."Like" WHERE "likedUser" = $1 AND "likingUser" = $2) AS visitedlikevisitor`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT exists(SELECT from public."Like" WHERE "likingUser" = $1 AND "likedUser" = $2) AS visitorlikevisited, exists(SELECT from public."Like" WHERE "likedUser" = $1 AND "likingUser" = $2) AS visitedlikevisitor`,
+        );
       const result = await db.any(
         `SELECT exists(SELECT from public."Like" WHERE "likingUser" = $1 AND "likedUser" = $2) AS visitorlikevisited, exists(SELECT from public."Like" WHERE "likedUser" = $1 AND "likingUser" = $2) AS visitedlikevisitor`,
         [visitorUser, visitedUser],
@@ -107,24 +117,27 @@ class Like {
         result[0].visitorlikevisited && result[0].visitedlikevisitor;
       return result[0];
     } catch (err) {
-      console.log(err, 'in model Like.exists()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.exists()');
       return null;
     }
   }
 
   async delete(likingUser, likedUser) {
     try {
-      console.log(
-        `DELETE FROM public."Like" WHERE "likingUser" = ${likingUser} AND "likedUser" = ${likedUser}`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `DELETE FROM public."Like" WHERE "likingUser" = ${likingUser} AND "likedUser" = ${likedUser}`,
+        );
       const result = await db.any(
         'DELETE FROM public."Like" WHERE "likingUser" = $1  AND "likedUser" = $2 RETURNING EXISTS(SELECT from public."Like" WHERE "likedUser" = $1 AND "likingUser" = $2) AS unmatch',
         [likingUser, likedUser],
       );
-      console.log(result[0].unmatch);
+      if (process.env.VERBOSE === 'true') console.log(result[0].unmatch);
       return { success: true, deleted: true, unmatch: result[0].unmatch };
     } catch (err) {
-      console.log(err, 'in model User.delete()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model User.delete()');
       return { deleted: false, error: err };
     }
   }
