+++ server/rest/components(C-M-R)/match/model.js
@@ -11,9 +11,10 @@ class Match {
 
   async create(user1, user2) {
     try {
-      console.log(
-        `INSERT INTO public."Match" (user1, user2, date) VALUES (${user1}, ${user2}, Now() RETURNING id`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `INSERT INTO public."Match" (user1, user2, date) VALUES (${user1}, ${user2}, Now() RETURNING id`,
+        );
       return await db
         .any(
           'INSERT INTO public."Match" ("user1", "user2", date) VALUES ($1, $2, NOW()) RETURNING id',
@@ -26,16 +27,18 @@ class Match {
           };
         });
     } catch (err) {
-      console.log(err, 'in model Like.create()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.create()');
       return { created: false, error: err };
     }
   }
 
   async updateLastMessage(matchId, messageId) {
     try {
-      console.log(
-        `UPDATE public."Match" SET "lastMessage" = ${messageId} WHERE id = ${matchId}`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `UPDATE public."Match" SET "lastMessage" = ${messageId} WHERE id = ${matchId}`,
+        );
       await db.any(
         'UPDATE public."Match" SET "lastMessage" = $2 WHERE id = $1',
         [matchId, messageId],
@@ -45,23 +48,26 @@ class Match {
         updated: true,
       };
     } catch (err) {
-      console.log(err, 'in model Like.create()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Like.create()');
       return { created: false, error: err };
     }
   }
 
   async delete(user1, user2) {
     try {
-      console.log(
-        `DELETE FROM public."Match" WHERE "user1" = ${user1} AND "user2" = ${user2}`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `DELETE FROM public."Match" WHERE "user1" = ${user1} AND "user2" = ${user2}`,
+        );
       await db.any(
         'DELETE FROM public."Match" WHERE ("user1" = $1  AND "user2" = $2) OR ("user1" = $2  AND "user2" = $1)',
         [user1, user2],
       );
       return { success: true, deleted: true };
     } catch (err) {
-      console.log(err, 'in model User.delete()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model User.delete()');
       return { deleted: false, error: err };
     }
   }
@@ -69,10 +75,12 @@ class Match {
   async getBy(type, value) {
     try {
       if (!this.isValidType(type)) {
-        console.log(`Match.getBy(): ${type} is not an authorized type`);
+        if (process.env.VERBOSE === 'true')
+          console.log(`Match.getBy(): ${type} is not an authorized type`);
         return null;
       }
-      console.log(`SELECT * FROM public."Match" WHERE ${type} = ${value}`);
+      if (process.env.VERBOSE === 'true')
+        console.log(`SELECT * FROM public."Match" WHERE ${type} = ${value}`);
       if (type instanceof Array && type.length === 2) {
         return db.any(
           `SELECT * FROM public."Match" WHERE $1:name = $3 OR $2:name = $3`,
@@ -85,16 +93,18 @@ class Match {
       );
       return result;
     } catch (err) {
-      console.log(err, 'in model Match.getBy()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Match.getBy()');
       return null;
     }
   }
 
   async getMatchId(user1, user2) {
     try {
-      console.log(
-        `SELECT id FROM public."Match" WHERE (user1 = ${user1} AND user2 = ${user2}) OR (user2 = ${user1} AND user1 = ${user2})`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT id FROM public."Match" WHERE (user1 = ${user1} AND user2 = ${user2}) OR (user2 = ${user1} AND user1 = ${user2})`,
+        );
       const result = await db.any(
         'SELECT id FROM public."Match" WHERE (user1 = $1 AND user2 = $2) OR (user2 = $1 AND user1 = $2)',
         [user1, user2],
@@ -102,16 +112,18 @@ class Match {
       if (result[0]) return result[0].id;
       else return null;
     } catch (err) {
-      console.log(err, 'in model Match.getAll()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Match.getAll()');
       return null;
     }
   }
 
   async getUsersFromMatchId(matchId) {
     try {
-      console.log(
-        `SELECT user1, user2 FROM public."Match" WHERE id = ${matchId})`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT user1, user2 FROM public."Match" WHERE id = ${matchId})`,
+        );
       const result = await db.any(
         'SELECT user1, user2 FROM public."Match" WHERE id = $1',
         [matchId],
@@ -121,18 +133,21 @@ class Match {
       }
       return false;
     } catch (err) {
-      console.log(err, 'in model Match.getAll()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Match.getAll()');
       return null;
     }
   }
 
   async getAll() {
     try {
-      console.log('SELECT * FROM public."Match"');
+      if (process.env.VERBOSE === 'true')
+        console.log('SELECT * FROM public."Match"');
       const result = await db.any('SELECT * FROM public."Match"');
       return result;
     } catch (err) {
-      console.log(err, 'in model Match.getAll()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Match.getAll()');
       return null;
     }
   }
@@ -141,19 +156,22 @@ class Match {
     try {
       if (!value) return false;
       if (!this.isValidType(type)) {
-        console.log(`Match.exists(): ${type} is not an authorized type`);
+        if (process.env.VERBOSE === 'true')
+          console.log(`Match.exists(): ${type} is not an authorized type`);
         return null;
       }
-      console.log(
-        `SELECT exists(SELECT from public."Match" WHERE ${type} = ${value})`,
-      );
+      if (process.env.VERBOSE === 'true')
+        console.log(
+          `SELECT exists(SELECT from public."Match" WHERE ${type} = ${value})`,
+        );
       const result = await db.any(
         `SELECT exists(SELECT from public."Match" WHERE id = $1);`,
         [value],
       );
       return result[0].exists;
     } catch (err) {
-      console.log(err, 'in model Match.exists()');
+      if (process.env.VERBOSE === 'true')
+        console.log(err, 'in model Match.exists()');
       return null;
     }
   }
